Move tooltip handler to client events package
