+++ src/main/java/dev/rndmorris/salisarcana/ClientProxy.java
@@ -9,9 +9,9 @@ import cpw.mods.fml.common.FMLCommonHandler;
 import cpw.mods.fml.common.event.FMLInitializationEvent;
 import cpw.mods.fml.common.event.FMLPostInitializationEvent;
 import dev.rndmorris.salisarcana.client.ThaumicInventoryScanner;
+import dev.rndmorris.salisarcana.client.events.WandPartTooltipEventHandler;
 import dev.rndmorris.salisarcana.client.handlers.GuiHandler;
 import dev.rndmorris.salisarcana.config.SalisConfig;
-import dev.rndmorris.salisarcana.lib.WandPartTooltipEventHandler;
 
 public class ClientProxy extends CommonProxy {
